Some minor refactoring
