+++ src/recoil/archive/selectors.ts
@@ -1,13 +1,14 @@
 import { selector, selectorFamily } from "recoil"
 import { ArchiveFileMetadata, RequiredFileInfo } from "../../types"
-import { RequiredFileInfoSelectorFamily } from "../selectors"
 import { PocketSyncConfigSelector } from "../config/selectors"
+import { RequiredFileInfoSelectorFamily } from "../requiredFiles/selectors"
 
 export const archiveMetadataSelector = selector<ArchiveFileMetadata[]>({
   key: "archiveMetadataSelector",
   get: async ({ get }) => {
     const config = get(PocketSyncConfigSelector)
     if (!config.archive_url) return []
+    // TODO: Support different archive sites
     const url = config.archive_url.replace("download", "metadata")
     const { files } = (await (await fetch(url)).json()) as {
       files: ArchiveFileMetadata[]
